Fix Share UI
resovles #652
